feature: Support directory and file creation by ec
ec is a shell script that supports directory and its corresponding file creation.
When we solve a problem in solved.ac, we need a python file and its
corresponding input file, and README markdown to save problem or make a comment on a problem.
To create each files on each problem, you may lose interest before you
solve problem.

ec script can be an option to give a nudge there.
It copies all the contents from directory `.sample` to the directory
given the first and the only argument to ec.
Of course, you can name the directory in Korean.

```example
./ec acimpc/problem_title
./ec acmipc/problem\ title\ with\ whitespace
./ec acimpc/한글제목
./ec another_platform/한글도\ 띄워쓰기는\ 백슬래시를
```